requirements.txt
Flask
psycopg2-binary
Werkzeug
Jinja2
itsdangerous
click
python-dotenv

